pytests/requirements.txt
PyGithub
base58
cython
deepdiff
ed25519
numpy
prometheus-client
pynacl
python-rc==0.3.9
requests
retrying
scikit-learn
# TODO(mina86): scipy 1.7.2 breaks buildkite so for the time being pin
# to an older version.  Remove the pin once issue is fixed properly.
scipy==1.7.1
semver
tqdm
